refactor: add keyword filtering to slack new-message trigger
